ISSUE #366 WIP
