+++ tools/bouncer_worker/src/lib/config.js
@@ -26,6 +26,8 @@ const applyDefaultValuesIfUndefined = (config) => {
 	config.rabbitmq.task_prefetch = config.rabbitmq.task_prefetch || 4;
 	config.rabbitmq.model_prefetch = config.rabbitmq.model_prefetch || 1;
 	config.bouncer.log_dir = config.bouncer.log_dir || ".log";
+	config.mongoimport = config.mongoimport || {};
+	config.mongoimport.writeConcern = config.mongoimport.writeConcern || { w: 1 };
 }
 
 const parseConfig = () => {
